+++ Makefile.am
@@ -564,8 +564,8 @@ ACLK_FILES = \
     aclk/aclk_rx_msgs.h \
     aclk/https_client.c \
     aclk/https_client.h \
-    aclk/aclk_charts.c \
-    aclk/aclk_charts.h \
+    aclk/aclk_charts_api.c \
+    aclk/aclk_charts_api.h \
     aclk/aclk_alarm_api.c \
     aclk/aclk_alarm_api.h \
     mqtt_websockets/src/mqtt_wss_client.c \
